+++ lib/src/flutter_advanced_networkimage.dart
@@ -104,9 +104,14 @@ class AdvancedNetworkImage extends ImageProvider<AdvancedNetworkImage> {
     String uId = uid(key.url);
 
     if (useDiskCache) {
-      Uint8List _diskCache = await _loadFromDiskCache(key, uId);
-      if (key.loadedCallback != null) key.loadedCallback();
-      return await PaintingBinding.instance.instantiateImageCodec(_diskCache);
+      try {
+        Uint8List _diskCache = await _loadFromDiskCache(key, uId);
+        if (key.loadedCallback != null) key.loadedCallback();
+        return await PaintingBinding.instance.instantiateImageCodec(_diskCache);
+      } catch (e) {
+        debugPrint(e.toString());
+        if (key.loadFailedCallback != null) key.loadFailedCallback();
+      }
     }
 
     Uint8List imageData = await _loadFromRemote(